feat(tickets): extract TicketDetailDrawer + paginated/filtered useTicketsList
- `src/components/tickets/ticket-detail-drawer.tsx` (new) — extracted
  drawer body from `TicketRow` so the hub's new Help Center
  `<TicketCard>` AND the embedded `<TicketCenter>` share the SAME
  composer + timeline + close/reopen + ModalV2 close confirmation.
  Owns: metadata strip, timeline panel (via `useTicketEngagements`),
  status-dependent action strip (text/attachment composer OR
  reopen-only). Local state for composer text + attachments + close
  dialog.
- `src/components/tickets/ticket-row.tsx` — slimmed from 547 → 89
  lines; now just wraps `<ChatTicketItem>` summary tile +
  `<TicketDetailDrawer>` in a `<Collapsible>`.
- `src/components/tickets/hooks/use-tickets-list.ts` — extends the
  filter surface with `page` + `pageSize` and surfaces `totalCount`
  / `totalPages` from the server's `{count:'exact'}` response.
  Cache key includes `[..., page, pageSize]` so each page/filter
  combo gets its own slot.
- `src/components/tickets/index.ts` — exports `TicketDetailDrawer`,
  `TicketRef`, `useTicketEngagements`, plus the engagement
  file/note types.
- `package.json` → 0.0.108
