+++ openframe-frontend-core/src/components/tickets/index.ts
@@ -1,8 +1,22 @@
 export { TicketCenter, type TicketCenterProps } from './ticket-center'
 export { TicketOpenForm, type TicketOpenFormProps } from './ticket-open-form'
 export { TicketRow, type TicketRowProps } from './ticket-row'
+export {
+  TicketDetailDrawer,
+  type TicketDetailDrawerProps,
+  type TicketRef,
+} from './ticket-detail-drawer'
 export { useTicketsList } from './hooks/use-tickets-list'
-export { useTicketActions, mapTicketActionError, type UseTicketActionsReturn } from './hooks/use-ticket-actions'
+export {
+  useTicketActions,
+  mapTicketActionError,
+  type UseTicketActionsReturn,
+} from './hooks/use-ticket-actions'
+export {
+  useTicketEngagements,
+  type TicketEngagement,
+  type TicketEngagementFile,
+} from './hooks/use-ticket-engagements'
 export {
   type TicketData,
   type OptimisticTicket,
